Add Gateway routing and define user-shelf-book relationships

+++ server/shelf-service/src/main/java/com/virtualshelf/shelf/service/ShelfService.java
@@ -1,14 +1,17 @@
 package com.virtualshelf.shelf.service;
 
 import com.virtualshelf.shelf.entity.Shelf;
+import com.virtualshelf.shelf.exception.ForbiddenException;
 import com.virtualshelf.shelf.repository.ShelfRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+
 @Service
 @RequiredArgsConstructor
 public class ShelfService {
+
     private final ShelfRepository shelfRepository;
 
     public Shelf addShelf(Shelf shelf) {
@@ -20,9 +23,13 @@ public class ShelfService {
                 .orElseThrow(() -> new RuntimeException("Shelf not found with id " + id));
     }
 
-    public Shelf updateShelf(Long id, Shelf updatedShelf) {
+    public Shelf updateShelf(Long id, Shelf updatedShelf, Long currentUserId) {
         Shelf existingShelf = shelfRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Shelf not found"));
+
+        if (!existingShelf.getUserId().equals(currentUserId)) {
+            throw new ForbiddenException("You are not authorized to update this shelf");
+        }
 
         if (updatedShelf.getName() != null) existingShelf.setName(updatedShelf.getName());
         if (updatedShelf.getIsPublic() != null) existingShelf.setIsPublic(updatedShelf.getIsPublic());
@@ -31,8 +38,14 @@ public class ShelfService {
     }
 
 
-    public void deleteShelf(Long id) {
+    public void deleteShelf(Long id, Long currentUserId) {
         Shelf existingShelf = getShelfById(id);
+
+        if (!existingShelf.getUserId().equals(currentUserId)) {
+            throw new ForbiddenException("You are not authorized to delete this shelf");
+        }
+
         shelfRepository.delete(existingShelf);
     }
 }
+
